Problem: median not accurate
Solution: use floats to determine if the middle
number falls within a *.5, if so add the numbers
together and divide by 2, else just return
the middle number.

+++ contracts/quadruple/contract.capnp
@@ -4,5 +4,5 @@ struct Quadruple {
   first @0 : UInt32;
   second @1 : UInt32;
   third @2 : UInt32;
-  fourth @3 : UInt32;
+  fourth @3 : Float32;
 }
